requirements.txt
# Core
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
openai==1.51.0

# Data processing & viz
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Web frameworks
Flask==3.0.0
Werkzeug==3.0.1
fastapi==0.110.0
uvicorn==0.27.0

# LLM helpers & parsing
beautifulsoup4==4.12.0
markdown2==2.4.0
ddgs==1.0.1

# Dev tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Export
playwright==1.49.0
